Amalgamated some coherence files, generalized t2star so that only one generating function needed, clean up tests
